refactor: move home screen callout type out of component file
